# performance: cache parsed markers, constraints and versions