--- a/src/poetry/core/version/pep440/parser.py
+++ b/src/poetry/core/version/pep440/parser.py
@@ -1,5 +1,6 @@
 from __future__ import annotations
 
+import functools
 import re
 
 from typing import TYPE_CHECKING
@@ -63,6 +64,7 @@ class PEP440Parser:
         )
 
     @classmethod
+    @functools.lru_cache(maxsize=None)
     def parse(cls, value: str, version_class: type[T]) -> T:
         match = cls._regex.search(value) if value else None
         if not match:
@@ -80,4 +82,4 @@ class PEP440Parser:
 
 
 def parse_pep440(value: str, version_class: type[T]) -> T:
-    return PEP440Parser.parse(value, version_class)
+    return PEP440Parser.parse(value, version_class)  # type: ignore[arg-type]