Some more fallout from #530. The parsing of marker strings became a bit slower with measurable effects in real-world examples. It seems in real-world examples, the same marker is used many times for different dependencies. Since markers are immutable, we can just cache the parsed markers and mitigate the performance regression.

Performance for shootout example with warm cache:

|command|1.3.2|master|PR (only markers)|PR (constraints, versions)|
|---|---|---|---|---|
|lock|46 s|57 s|47 s|35 s|
|lock --no-update|14 s|18 s|14 s|10 s|

(There seems to be no significant increase in peak memory usage due to the caching.)

Update: added times for caching parsed constraints and versions in addition to markers.