@dimbleby 2023-02-19
looks fine to me though fwiw I don't reproduce any performance delta on the "big" install that I have lying around.