@dimbleby 2023-02-19
if this is real: I wonder whether constraint parsing and version parsing would similarly benefit